clean up resources

+++ enjoy/src/main/db/models/audio.ts
@@ -258,7 +258,9 @@ export class Audio extends Model<Audio> {
 
   @AfterDestroy
   static cleanupFile(audio: Audio) {
-    fs.remove(audio.filePath);
+    if (audio.filePath) {
+      fs.remove(audio.filePath);
+    }
     Recording.destroy({
       where: {
         targetId: audio.id,
